Fix display error, refs #50
When the popup was disappearing, the X would show outside the popup for
a moment. This is caused by React-Leaflet(?) removing the popup content
while it is being removed, so the popup becomes empty during the exit
animation. That causes the X to not be where it was when the popup was
open. This commit just removes the animation.